cleaning up and organizing since example now working
