+++ applications/variational/examples/FCC/fcc.out
@@ -1,0 +1,26 @@
+0 1 1
+0.117507 1.02938 1.003
+0.177048 1.04425 1.00452
+0.236594 1.0591 1.00604
+0.296148 1.07393 1.00757
+0.355706 1.08873 1.0091
+0.415226 1.10369 1.01064
+0.474565 1.11934 1.01218
+0.533384 1.13686 1.01369
+0.591036 1.15789 1.01508
+0.646586 1.18396 1.01625
+0.699359 1.21527 1.01704
+0.750015 1.24994 1.01733
+0.800672 1.2846 1.01705
+0.853448 1.31591 1.01626
+0.908999 1.34197 1.0151
+0.966652 1.36301 1.01371
+1.02547 1.38053 1.0122
+1.0848 1.3962 1.01065
+1.14432 1.41117 1.00911
+1.20387 1.42599 1.00758
+1.26342 1.44083 1.00605
+1.32296 1.4557 1.00453
+1.3825 1.47059 1.003
+1.44204 1.48549 1.00148
+1.5 1.5 1
