+++ applications/variational/examples/_lj.gfg
@@ -1,0 +1,2 @@
+0.212469	0.580506	1.796850	1.0	1.0
+2.272498	0.502913	1.765567	1.0	1.0
