+++ applications/variational/examples/lj_atoms/lj_atoms.pov
@@ -1,0 +1,31 @@
+// begin gfg2pov records
+sphere{<1.695601, 2.466513, 1.806581>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<2.032058, 2.287045, 2.693562>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<1.616258, 1.572390, 1.905550>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<0.901778, 2.039993, 1.559657>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<0.387574, 0.539868, 2.631421>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<2.219364, 0.154657, 1.090029>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<1.763306, 0.624605, 1.791073>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<2.679959, 0.787321, 0.547327>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<1.184597, 2.798837, 2.783740>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<2.406742, 1.046981, 2.344262>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<0.469138, 2.357308, 2.431246>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<0.630972, 2.136894, 0.604892>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<2.501461, 0.043815, 2.159225>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<0.763703, 1.152005, 0.488637>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<2.031094, 0.365088, 0.051108>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<0.104052, 0.708066, 1.590667>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<2.795437, 2.702772, 0.322154>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<2.368023, 2.071714, 0.921043>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<1.306719, 0.887801, 2.560256>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<1.918138, 1.368140, 0.309144>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<0.991535, 1.078232, 1.406811>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<0.132958, 2.560871, 1.409943>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<2.826906, 1.704427, 0.137070>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<0.539452, 1.417863, 2.193145>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<1.219813, 1.858926, 2.730195>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<2.091891, 1.233134, 1.213904>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<0.214183, 1.534565, 1.165752>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<1.504156, 2.459042, 0.850899>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<1.490914, 0.606519, 0.766707>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
+sphere{<0.899523, 0.174787, 1.811784>, 0.500000 texture{ pigment {color White  transmit 0.800000  }  finish {phong 1.000000}  } }
