+++ applications/variational/examples/ljx.gfg
@@ -1,0 +1,32 @@
+1.912498	1.498022	1.757718	1.0	1.0
+2.086439	1.104615	3.087084	1.0	1.0
+1.222874	0.252912	1.698112	1.0	1.0
+1.467486	0.322127	2.881935	1.0	1.0
+1.005686	1.318325	0.145087	1.0	1.0
+1.059963	1.075784	2.323194	1.0	1.0
+0.438524	0.429635	2.829934	1.0	1.0
+0.212469	0.580506	1.796850	1.0	1.0
+2.272498	0.502913	1.765567	1.0	1.0
+2.761660	1.329101	1.198805	1.0	1.0
+2.513476	0.218820	2.986670	1.0	1.0
+2.818715	0.321718	0.778860	1.0	1.0
+1.286187	2.580505	0.256872	1.0	1.0
+2.993843	2.627819	0.212736	1.0	1.0
+2.392704	2.139965	2.830647	1.0	1.0
+0.472119	1.761644	1.791283	1.0	1.0
+2.591445	2.336077	1.394699	1.0	1.0
+3.035464	2.680453	2.247329	1.0	1.0
+1.943918	1.883894	0.739796	1.0	1.0
+1.461124	2.002994	2.654747	1.0	1.0
+0.735977	0.417035	0.672415	1.0	1.0
+1.928527	2.847795	2.249184	1.0	1.0
+0.839469	2.775709	2.531437	1.0	1.0
+0.319170	1.901374	2.801081	1.0	1.0
+3.174685	1.138164	0.215135	1.0	1.0
+0.939024	1.207051	1.182912	1.0	1.0
+0.349739	2.004931	0.726174	1.0	1.0
+0.451489	2.806422	1.353870	1.0	1.0
+2.862317	1.215953	2.330381	1.0	1.0
+1.857373	0.846424	0.967600	1.0	1.0
+2.079102	2.921497	0.686538	1.0	1.0
+1.414659	2.536554	1.329974	1.0	1.0
