+++ applications/variational/examples/random/rando.gfg
@@ -1,0 +1,10 @@
+0.6656770517456656	0.5446433993113982	0.0023286810248260847	0.25	1.0
+0.456162983355769	0.11509471353087242	0.4463230858822308	0.25	1.0
+0.8131141019338464	0.31483777691961956	0.9463775164864228	0.25	1.0
+0.2438717543531328	0.4083518549027747	0.48873956310610345	0.25	1.0
+0.21657223285664395	0.17013943194401315	0.7935653935435382	0.25	1.0
+0.011578153135018088	0.65999623597165	0.7558228767407297	0.25	1.0
+0.25573589347846226	0.996236502364124	0.21679067690728837	0.25	1.0
+0.01944556548848042	0.5774199299193967	0.3461078742168885	0.25	1.0
+0.6283393344048422	0.3533471649047112	0.4091112310448559	0.25	1.0
+0.08514625700441736	0.1571909930240466	0.22142322650645263	0.25	1.0
