+++ applications/variational/examples/random/rando2.gfg
@@ -1,0 +1,80 @@
+0.665677	0.544643	0.002329	0.25	1.0
+0.665677	0.544643	1.002329	0.25	1.0
+0.665677	1.544643	0.002329	0.25	1.0
+0.665677	1.544643	1.002329	0.25	1.0
+1.665677	0.544643	0.002329	0.25	1.0
+1.665677	0.544643	1.002329	0.25	1.0
+1.665677	1.544643	0.002329	0.25	1.0
+1.665677	1.544643	1.002329	0.25	1.0
+0.456163	0.115095	0.446323	0.25	1.0
+0.456163	0.115095	1.446323	0.25	1.0
+0.456163	1.115095	0.446323	0.25	1.0
+0.456163	1.115095	1.446323	0.25	1.0
+1.456163	0.115095	0.446323	0.25	1.0
+1.456163	0.115095	1.446323	0.25	1.0
+1.456163	1.115095	0.446323	0.25	1.0
+1.456163	1.115095	1.446323	0.25	1.0
+0.813114	0.314838	0.946378	0.25	1.0
+0.813114	0.314838	1.946378	0.25	1.0
+0.813114	1.314838	0.946378	0.25	1.0
+0.813114	1.314838	1.946378	0.25	1.0
+1.813114	0.314838	0.946378	0.25	1.0
+1.813114	0.314838	1.946378	0.25	1.0
+1.813114	1.314838	0.946378	0.25	1.0
+1.813114	1.314838	1.946378	0.25	1.0
+0.243872	0.408352	0.488740	0.25	1.0
+0.243872	0.408352	1.488740	0.25	1.0
+0.243872	1.408352	0.488740	0.25	1.0
+0.243872	1.408352	1.488740	0.25	1.0
+1.243872	0.408352	0.488740	0.25	1.0
+1.243872	0.408352	1.488740	0.25	1.0
+1.243872	1.408352	0.488740	0.25	1.0
+1.243872	1.408352	1.488740	0.25	1.0
+0.216572	0.170139	0.793565	0.25	1.0
+0.216572	0.170139	1.793565	0.25	1.0
+0.216572	1.170139	0.793565	0.25	1.0
+0.216572	1.170139	1.793565	0.25	1.0
+1.216572	0.170139	0.793565	0.25	1.0
+1.216572	0.170139	1.793565	0.25	1.0
+1.216572	1.170139	0.793565	0.25	1.0
+1.216572	1.170139	1.793565	0.25	1.0
+0.011578	0.659996	0.755823	0.25	1.0
+0.011578	0.659996	1.755823	0.25	1.0
+0.011578	1.659996	0.755823	0.25	1.0
+0.011578	1.659996	1.755823	0.25	1.0
+1.011578	0.659996	0.755823	0.25	1.0
+1.011578	0.659996	1.755823	0.25	1.0
+1.011578	1.659996	0.755823	0.25	1.0
+1.011578	1.659996	1.755823	0.25	1.0
+0.255736	0.996237	0.216791	0.25	1.0
+0.255736	0.996237	1.216791	0.25	1.0
+0.255736	1.996237	0.216791	0.25	1.0
+0.255736	1.996237	1.216791	0.25	1.0
+1.255736	0.996237	0.216791	0.25	1.0
+1.255736	0.996237	1.216791	0.25	1.0
+1.255736	1.996237	0.216791	0.25	1.0
+1.255736	1.996237	1.216791	0.25	1.0
+0.019446	0.577420	0.346108	0.25	1.0
+0.019446	0.577420	1.346108	0.25	1.0
+0.019446	1.577420	0.346108	0.25	1.0
+0.019446	1.577420	1.346108	0.25	1.0
+1.019446	0.577420	0.346108	0.25	1.0
+1.019446	0.577420	1.346108	0.25	1.0
+1.019446	1.577420	0.346108	0.25	1.0
+1.019446	1.577420	1.346108	0.25	1.0
+0.628339	0.353347	0.409111	0.25	1.0
+0.628339	0.353347	1.409111	0.25	1.0
+0.628339	1.353347	0.409111	0.25	1.0
+0.628339	1.353347	1.409111	0.25	1.0
+1.628339	0.353347	0.409111	0.25	1.0
+1.628339	0.353347	1.409111	0.25	1.0
+1.628339	1.353347	0.409111	0.25	1.0
+1.628339	1.353347	1.409111	0.25	1.0
+0.085146	0.157191	0.221423	0.25	1.0
+0.085146	0.157191	1.221423	0.25	1.0
+0.085146	1.157191	0.221423	0.25	1.0
+0.085146	1.157191	1.221423	0.25	1.0
+1.085146	0.157191	0.221423	0.25	1.0
+1.085146	0.157191	1.221423	0.25	1.0
+1.085146	1.157191	0.221423	0.25	1.0
+1.085146	1.157191	1.221423	0.25	1.0
