+++ applications/variational/examples/random/random.gfg
@@ -1,0 +1,10 @@
+0.07084139675070888	0.31544701273107834	0.04704332520053034	0.25	1.0
+0.3665317321742758	0.31395248622853433	0.12020064777911821	0.25	1.0
+0.6329283862515473	0.8315396961219331	0.859112740758025	0.25	1.0
+0.8923089852298893	0.983410496102354	0.16581814032245723	0.25	1.0
+0.25983265313230675	0.7373623096801093	0.2721472498331289	0.25	1.0
+0.32111594392663456	0.7321656533567406	0.46805257479811735	0.25	1.0
+0.3585804873207554	0.3650313544917252	0.6363437988850651	0.25	1.0
+0.6131446559853279	0.3214321671077979	0.1174915630465928	0.25	1.0
+0.8110097764507455	0.3223949822382577	0.25027750773409874	0.25	1.0
+0.6828155501122395	0.8819464760090482	0.6583414136883909	0.25	1.0
